+++ langserver/document.js
@@ -4,7 +4,7 @@ const os = require('os');
 const { CEIType } = require('java-mti');
 const { Settings } = require('./settings');
 const { ParseProblem } = require('./java/parser');
-const { parse } = require('./java/body-parser3');
+const { parse } = require('./java/body-parser');
 const { SourceUnit } = require('./java/source-types');
 const { parseMethodBodies } = require('./java/validater');
 const { time, timeEnd, trace } = require('./logging');
